refactor(room-layout): migrate to immutable layout snapshot model
Scope: room-layout, showtime, showtime-seat, exception

**Problem**
- Old model stored seats directly (room → seat), making any seat type/status change affect all past and future showtimes.
- No versioning or history of layout modifications, causing data inconsistency and operational risks.

**Solution**
Replace direct seat table with snapshot-based `RoomLayout` and `RoomLayoutSeat`:
- `RoomLayout` holds version, effective date, and dimensions.
- `RoomLayoutSeat` stores seat position, type, status, and couple group.
- Each layout change creates a new version (immutable).
- `Showtime` references a specific `room_layout_id`.
- `ShowtimeSeat` links to `room_layout_seat_id` and snapshots the ticket price at creation.

**Detailed changes**

1. **Domain & Entities (room-layout)**
   - Remove `Seat` domain and JPA entity (package `seat` deleted).
   - Add `RoomLayout` (aggregate root) and `RoomLayoutSeat`.
   - Add `room_layout_id` to `Showtime`; remove direct seat relationship.
   - Add `price` field to `ShowtimeSeat` (snapshot of ticket price).

2. **Repositories**
   - `RoomLayoutRepository` (only `create`, no update) with methods:
     - `findCurrentByRoomIdAndDate`, `findLatestByRoomId`, `findByRoomIdAndLayoutVersion`.
   - `RoomLayoutSeatRepository` (read-only queries).
   - Update `ShowtimeSeatRepository` to use `room_layout_seat_id`.

3. **Use Cases (application layer)**
   - `GenerateRoomLayoutUseCase`: initial layout version 1.
   - `UpdateRoomLayoutSeatsUseCase`: bulk seat type/status changes → creates new layout version.
   - `GetRoomLayoutUseCase`: returns current layout (highest version with effectiveDate ≤ today).
   - Delete all old seat update use cases (`BulkActiveSeatUseCase`, `BulkUpdateSeatTypeUseCase`, `DeleteSeatUsecase`, etc.).
   - Update `CreateShowtimeUsecase` to fetch current layout and generate `ShowtimeSeat` with price snapshot.

4. **Controllers & API**
   - Remove `SeatController` entirely.
   - Add endpoints in `RoomController`:
     - `POST /api/v1/rooms/{id}/layouts/update-seats` (admin) – create new layout version.
     - `GET /api/v1/rooms/{id}/layout` – get current layout.
   - Add `GET /api/v1/showtimes/{id}/seats` in `ShowtimePublicController` to display seat map.

5. **Error Handling & Exceptions**
   - Add `RoomLayoutExceptions` with dedicated error codes (3034–3043).
   - Extend `ErrorCode` enum for layout and layout seat errors.
   - Use `@PreAuthorize("hasRole('ADMIN')")` for mutation endpoints.
   - Enhance validation in all use cases (null checks, business rules).

6. **Database Schema (migration)**
   - Create `room_layouts` and `room_layout_seats`.
   - Add `room_layout_id` to `showtimes`.
   - Change `showtime_seats.seat_id` to reference `room_layout_seats.id`.
   - Add `price` column to `showtime_seats`.
   - (Migration script for existing data provided separately)

**Impact & backward compatibility**
- All old APIs depending on `seat` table are removed.
- Clients must switch to new layout APIs.
- Past showtimes remain unchanged after layout updates.
- Existing production data must be migrated (script provided).

**Testing**
- Unit/integration tests for layout versioning, showtime creation, and seat map retrieval.
- Verified that showtimes created before a layout change still use the old layout.

Ref: #CIN-123 (layout snapshot), #CIN-124 (showtime adaptation)

+++ backend/src/main/java/com/cinemaebooking/backend/showtime/application/dto/showtimeFormat/UpdateShowtimeFormatRequest.java
@@ -3,9 +3,11 @@ package com.cinemaebooking.backend.showtime.application.dto.showtimeFormat;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 public class UpdateShowtimeFormatRequest {
     private String name;
-    private Long extraPrice;
+    private BigDecimal extraPrice;
 }
